ppfinjector#40: Add SimplePatcher
* Adds IPatcher interface to support runtime patcher selection.
* Adds a simplified patcher to replace FlatPatch.

+++ tk/ppftk/inc/ppftk/rom_patch/patch_item.h
@@ -9,7 +9,7 @@ namespace tdd::tk::rompatch {
 
    struct [[nodiscard]] PatchItem
    {
-      size_t address;
+      uint64_t address;
       DataBuffer data;
 
       [[nodiscard]] bool operator<(const PatchItem& other) const noexcept
